int型のオーバーフロー問題を修正
問題:
- Lichess Opening Explorer APIが返す対局数が35億を超える場合がある
- int型の最大値(21億)を超えてNumberFormatExceptionが発生

修正内容:
- OpeningResponse.javaのwhite, draws, blackをintからlongに変更
- OpeningMove.javaのwhite, draws, blackをintからlongに変更
- getTotalGames()の戻り値もlongに変更

これで最大9京までの対局数に対応可能

+++ src/main/java/jp/ac/dendai/model/OpeningResponse.java
@@ -3,20 +3,20 @@ package jp.ac.dendai.model;
 import java.util.List;
 
 public class OpeningResponse {
-    private int white;
-    private int draws;
-    private int black;
+    private long white;
+    private long draws;
+    private long black;
     private List<OpeningMove> moves;
-    
-    public int getWhite() { return white; }
-    public void setWhite(int white) { this.white = white; }
-    
-    public int getDraws() { return draws; }
-    public void setDraws(int draws) { this.draws = draws; }
-    
-    public int getBlack() { return black; }
-    public void setBlack(int black) { this.black = black; }
-    
+
+    public long getWhite() { return white; }
+    public void setWhite(long white) { this.white = white; }
+
+    public long getDraws() { return draws; }
+    public void setDraws(long draws) { this.draws = draws; }
+
+    public long getBlack() { return black; }
+    public void setBlack(long black) { this.black = black; }
+
     public List<OpeningMove> getMoves() { return moves; }
     public void setMoves(List<OpeningMove> moves) { this.moves = moves; }
 }
